Split endpoint module into package (#27)
* Split the endpoint module

* Use strings in __all__

* Clean up docstring for StubEndpoint

* Simplify endpoint filenames

* Prep for version 2.2.0

* Add fixed section about wheel dep
