+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2019-02-04
+### Added
+- Automated release artifact deployment via Travis CI's `providers` feature
+
+### Changed
+- The `endpoint` module has been split up into a module for each endpoint type, still importable as before
+
+### Fixed
+- Added the missing `wheel` dependency to `dev-requirements.txt` for creating a wheel distribution during release
+
 ## [2.1.0] - 2019-01-31
 ### Added
 - Added ability to create a `StubEndpoint` for stubs
